Reorganizes and adds comments. General clean up
