+++ src/components/reroll/library/images/forms/index.ts
@@ -1,2 +1,6 @@
+export * from "components/reroll/library/images/forms/ImageCreationForm";
+export * from "components/reroll/library/images/forms/ImageForm";
+export * from "components/reroll/library/images/forms/ImageSelectionForm";
 export * from "components/reroll/library/images/forms/LinkImage";
 export * from "components/reroll/library/images/forms/UploadImage";
+
